Gating plot working. Still needs documentation cleanup. Graph needs to reset zoom on transform change. Color mapping broken.
