+++ src/app/constants.js
@@ -16,6 +16,7 @@ export const Component = {
   GENOMIC_PROFILES: 'genomicProfiles',
   CELL_SET_EXPRESSION: 'cellSetExpression',
   EXPRESSION_HISTOGRAM: 'expressionHistogram',
+  GATING: 'gating',
 };
 
 export const DataType = {
