+++ src/app/state/coordination.js
@@ -121,6 +121,36 @@ export const COMPONENT_COORDINATION_TYPES = {
     CoordinationType.OBS_COLOR_ENCODING,
     CoordinationType.ADDITIONAL_OBS_SETS,
   ],
+  [Component.GATING]: [
+    CoordinationType.DATASET,
+    CoordinationType.OBS_TYPE,
+    CoordinationType.FEATURE_TYPE,
+    CoordinationType.FEATURE_VALUE_TYPE,
+    CoordinationType.EMBEDDING_TYPE,
+    CoordinationType.EMBEDDING_ZOOM,
+    CoordinationType.EMBEDDING_ROTATION,
+    CoordinationType.EMBEDDING_TARGET_X,
+    CoordinationType.EMBEDDING_TARGET_Y,
+    CoordinationType.EMBEDDING_TARGET_Z,
+    CoordinationType.EMBEDDING_OBS_SET_POLYGONS_VISIBLE,
+    CoordinationType.EMBEDDING_OBS_SET_LABELS_VISIBLE,
+    CoordinationType.EMBEDDING_OBS_SET_LABEL_SIZE,
+    CoordinationType.EMBEDDING_OBS_RADIUS,
+    CoordinationType.EMBEDDING_OBS_RADIUS_MODE,
+    CoordinationType.EMBEDDING_OBS_OPACITY,
+    CoordinationType.EMBEDDING_OBS_OPACITY_MODE,
+    CoordinationType.OBS_FILTER,
+    CoordinationType.OBS_HIGHLIGHT,
+    CoordinationType.OBS_SET_SELECTION,
+    CoordinationType.OBS_SET_HIGHLIGHT,
+    CoordinationType.OBS_SET_COLOR,
+    CoordinationType.FEATURE_HIGHLIGHT,
+    CoordinationType.FEATURE_SELECTION,
+    CoordinationType.FEATURE_VALUE_COLORMAP,
+    CoordinationType.FEATURE_VALUE_COLORMAP_RANGE,
+    CoordinationType.OBS_COLOR_ENCODING,
+    CoordinationType.ADDITIONAL_OBS_SETS,
+  ],
   [Component.SPATIAL]: [
     CoordinationType.DATASET,
     CoordinationType.OBS_TYPE,
