+++ src/components/gating/GatingSubscriber.js
@@ -1,0 +1,181 @@
+import React, {
+  useMemo, useEffect,
+} from 'react';
+import BaseScatterplotSubscriber, {
+  BASE_SCATTERPLOT_DATA_TYPES,
+} from '../base-scatterplot/BaseScatterplotSubscriber';
+import {
+  useLoaders,
+  useCoordination,
+} from '../../app/state/hooks';
+import {
+  useReady, useUrls,
+} from '../hooks';
+import {
+  useCellsData,
+  useExpressionMatrixData,
+} from '../data-hooks';
+import OptionSelect from '../shared-plot-options/OptionSelect';
+import { useStyles } from '../shared-plot-options/styles';
+import { COMPONENT_COORDINATION_TYPES } from '../../app/state/coordination';
+
+const GATING_MAPPING_NAME = 'GATING';
+
+export default function GatingSubscriber(props) {
+  const {
+    coordinationScopes,
+  } = props;
+
+  const [urls, addUrl, resetUrls] = useUrls();
+  const loaders = useLoaders();
+  const [isReady, setItemIsReady, setItemIsNotReady, resetReadyItems] = useReady(
+    BASE_SCATTERPLOT_DATA_TYPES,
+  );
+
+  // Get "props" from the coordination space.
+  const [{ dataset }] = useCoordination(
+    COMPONENT_COORDINATION_TYPES.scatterplot,
+    coordinationScopes,
+  );
+
+  // Reset file URLs and loader progress when the dataset has changed.
+  useEffect(() => {
+    resetUrls();
+    resetReadyItems();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [loaders, dataset]);
+
+  const [selectedGenes, setSelectedGenes] = React.useState([]);
+  const TRANSFORM_OPTIONS = ['', 'Log', 'ArcSinh'];
+  const [transformType, setTransformType] = React.useState('');
+
+  const [cells, cellsCount] = useCellsData(loaders, dataset, setItemIsReady, addUrl, true);
+  const [expressionMatrix] = useExpressionMatrixData(
+    loaders, dataset, setItemIsReady, addUrl, true,
+  );
+
+  const geneSelectOptions = expressionMatrix && expressionMatrix.cols ? expressionMatrix.cols : [];
+
+  const handleGeneSelectChange = (event) => {
+    const { options } = event.target;
+    const newValues = [];
+    for (let i = 0, l = options.length; i < l; i += 1) {
+      if (options[i].selected) {
+        newValues.push(options[i].value);
+      }
+    }
+    if (newValues.length === 1 && selectedGenes.length === 1 && newValues[0] !== selectedGenes[0]) {
+      newValues.unshift(selectedGenes[0]);
+    }
+    if (newValues.length <= 2) setSelectedGenes(newValues);
+  };
+
+  const handleTransformSelectChange = (event) => {
+    setTransformType(event.target.value);
+  };
+
+  const classes = useStyles();
+  const customOptions = [
+    {
+      label: 'Genes',
+      input: (
+        <OptionSelect
+          key="gating-gene-select"
+          multiple
+          className={classes.select}
+          value={selectedGenes}
+          onChange={handleGeneSelectChange}
+          inputProps={{
+            id: 'scatterplot-gene-select',
+          }}
+        >
+          {geneSelectOptions.map(name => (
+            <option key={name} value={name}>
+              {name}
+            </option>
+          ))}
+        </OptionSelect>
+      ),
+    },
+    {
+      label: 'Transform',
+      input: (
+        <OptionSelect
+          key="gating-transform-select"
+          className={classes.select}
+          value={transformType}
+          onChange={handleTransformSelectChange}
+          inputProps={{
+            id: 'scatterplot-transform-select',
+          }}
+        >
+          {TRANSFORM_OPTIONS.map(name => (
+            <option key={name} value={name}>
+              {name}
+            </option>
+          ))}
+        </OptionSelect>
+      ),
+    },
+  ];
+
+  const cellsWithGenes = useMemo(
+    () => {
+      if (selectedGenes.length < 2) {
+        return [];
+      }
+
+      let transformFunction;
+      switch (transformType) {
+        case 'Log':
+          transformFunction = v => Math.log10(v + 1);
+          break;
+        case 'ArcSinh':
+          transformFunction = v => Math.asinh(v);
+          break;
+        default:
+          transformFunction = v => v;
+      }
+
+      const selectedGeneCols = [
+        expressionMatrix.cols.indexOf(selectedGenes[0]),
+        expressionMatrix.cols.indexOf(selectedGenes[1]),
+      ];
+      const updatedCells = {};
+      expressionMatrix.rows.forEach((cellId, index) => {
+        const curCell = cells[cellId];
+        const cellMatrixRowOffset = expressionMatrix.cols.length * index;
+        curCell.mappings[GATING_MAPPING_NAME] = [
+          transformFunction(expressionMatrix.matrix[cellMatrixRowOffset + selectedGeneCols[0]]),
+          transformFunction(expressionMatrix.matrix[cellMatrixRowOffset + selectedGeneCols[1]]),
+        ];
+        updatedCells[cellId] = curCell;
+      });
+      return updatedCells;
+    },
+    [cells, expressionMatrix, selectedGenes, transformType],
+  );
+
+  const title = useMemo(
+    () => {
+      if (selectedGenes.length < 2) {
+        return 'Gating';
+      }
+      return `Gating (${selectedGenes[0]} vs ${selectedGenes[1]})`;
+    }, [selectedGenes],
+  );
+
+  return (
+    <BaseScatterplotSubscriber
+      {...props}
+      loaders={loaders}
+      cellsData={[cellsWithGenes, cellsCount]}
+      expressionMatrixOrAttrs={expressionMatrix}
+      useReadyData={[isReady, setItemIsReady, setItemIsNotReady, resetReadyItems]}
+      urlsData={[urls, addUrl, resetUrls]}
+      mapping={GATING_MAPPING_NAME}
+      title={title}
+      customOptions={customOptions}
+    />
+  );
+}
