+++ src/components/gating/index.js
@@ -1,0 +1,1 @@
+export { default as GatingSubscriber } from './GatingSubscriber';
